FileReceiver/db.py
# db.py
import os
from sqlmodel import SQLModel
from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine, async_sessionmaker
from dotenv import load_dotenv
from sqlmodel.ext.asyncio.session import AsyncSession

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")

engine: AsyncEngine = create_async_engine(DATABASE_URL, echo=True)


async_session = async_sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)
