FileReceiver/requirements.txt
aiofiles==24.1.0
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
databases==0.9.0
decorator==5.2.1
dnspython==2.7.0
dogpile.cache==1.4.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kafka-python==2.0.5
keystoneauth1==5.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openstacksdk==4.5.0
os-service-types==1.7.0
packaging==25.0
pbr==6.1.1
peewee==3.17.9
platformdirs==4.3.8
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requestsexceptions==1.4.0
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.0
stevedore==5.4.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
sqlmodel~=0.0.24